DeepLearning_Report/requirements.txt
matplotlib==3.7.1
mediapipe==0.10.0
numpy==1.23.5
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
Pillow==9.4.0
Pillow==9.5.0
tensorboardX==2.6
tensorboardX==2.6
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
